Further Cleanup & Gradle
I added the Gradle stuff cuz originally there wasn't any Maven or Gradle in the source folder and I guess it would be helpful for other ppl looking to contribute and already added the needed libs in src/libs.
