--- ReduceSneakDMG/.classpath
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
-		<attributes>
-			<attribute name="module" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="C:/Users/Zach/Desktop/BuildTools/spigot-1.17.jar"/>
-	<classpathentry kind="lib" path="C:/Users/Zach/Desktop/ActionAnnouncer.jar"/>
-	<classpathentry kind="lib" path="C:/Users/Zach/eclipse-workspace/worldedit-bukkit-7.2.0-SNAPSHOT-dist.jar"/>
-	<classpathentry kind="lib" path="C:/Users/Zach/eclipse-workspace/worldguard-bukkit-7.0.4-SNAPSHOT-dist.jar"/>
-	<classpathentry kind="lib" path="C:/Users/Zach/eclipse-workspace/EssentialsX-2.19.0-dev+85-e5db842.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
