+++ ReduceSneakDMG/gradle.properties
@@ -1,0 +1,3 @@
+org.gradle.jvmargs=-Xmx3G
+org.gradle.caching=true
+org.gradle.parallel=true
